chore: move around code
