+++ src/rudra/analysis/send_sync_variance/analyze.rs
@@ -1,61 +1,15 @@
 use super::{
-    utils::{self_type, AdtGenericParams, Krate, TraitDid},
+    utils::{self_type, AdtGenericParams, TraitDid},
     Tag,
 };
-use crate::rudra::context::RudraCtxt;
 use charon_lib::{
     ast::{TraitImpl, TranslatedCrate, Ty, TypeVarId},
-    formatter::{FmtCtx, IntoFormatter},
+    formatter::FmtCtx,
     ids::Vector,
     pretty::FmtWithCtx,
 };
 
-pub struct SendSyncChecker<'tcx> {
-    rcx: RudraCtxt<'tcx>,
-    trait_dids: TraitDid,
-    send_impls: Vec<&'tcx TraitImpl>,
-    sync_impls: Vec<&'tcx TraitImpl>,
-}
-
-impl<'tcx> SendSyncChecker<'tcx> {
-    pub fn new(rcx: RudraCtxt<'tcx>) -> Self {
-        let krate = Krate::new(&rcx.crate_data);
-        let trait_dids = krate.send_sync_copy();
-        Self {
-            rcx,
-            send_impls: krate.trait_impls(trait_dids.send),
-            sync_impls: krate.trait_impls(trait_dids.sync),
-            trait_dids,
-        }
-    }
-
-    pub fn analyze(self) {
-        let krate = &self.rcx.crate_data;
-        let ctx = &krate.into_fmt();
-
-        for imp in &self.send_impls {
-            analyze_send(imp, &self.trait_dids, krate, ctx);
-        }
-
-        for imp in &self.sync_impls {
-            analyze_sync(imp, &self.trait_dids, krate, ctx);
-        }
-    }
-}
-
-fn generic_params(imp: &TraitImpl, krate: &TranslatedCrate, ctx: &FmtCtx) -> AdtGenericParams {
-    let this = self_type(imp, ctx);
-    let Some(adt) = this.as_adt().and_then(|t| t.0.as_adt()) else {
-        panic!(
-            "Display:{}\nDebug:{0:?}\nin send impl should be an adt.",
-            this.fmt_with_ctx(ctx),
-        );
-    };
-
-    AdtGenericParams::new(krate, *adt)
-}
-
-fn analyze_send(imp: &TraitImpl, traits: &TraitDid, krate: &TranslatedCrate, fmt: &FmtCtx) {
+pub fn analyze_send(imp: &TraitImpl, traits: &TraitDid, krate: &TranslatedCrate, fmt: &FmtCtx) {
     let mut adt_type_params = generic_params(imp, krate, fmt);
     adt_type_params.add_trait_bounds_on_impl(imp, fmt, true);
 
@@ -85,7 +39,7 @@ fn analyze_send(imp: &TraitImpl, traits: &TraitDid, krate: &TranslatedCrate, fmt
     }
 }
 
-fn analyze_sync(imp: &TraitImpl, traits: &TraitDid, krate: &TranslatedCrate, fmt: &FmtCtx) {
+pub fn analyze_sync(imp: &TraitImpl, traits: &TraitDid, krate: &TranslatedCrate, fmt: &FmtCtx) {
     let mut adt_type_params = generic_params(imp, krate, fmt);
     adt_type_params.add_trait_bounds_on_impl(imp, fmt, false);
 
@@ -171,4 +125,16 @@ impl<'a, 'b> TraitImplCxt<'a, 'b> {
             source_text.unwrap_or_else(|| self.imp.fmt_with_ctx(self.fmt))
         })
     }
+}
+
+fn generic_params(imp: &TraitImpl, krate: &TranslatedCrate, ctx: &FmtCtx) -> AdtGenericParams {
+    let this = self_type(imp, ctx);
+    let Some(adt) = this.as_adt().and_then(|t| t.0.as_adt()) else {
+        panic!(
+            "Display:{}\nDebug:{0:?}\nin send impl should be an adt.",
+            this.fmt_with_ctx(ctx),
+        );
+    };
+
+    AdtGenericParams::new(krate, *adt)
 }
